fix: refactor Task_Update class
moved to shared folder as now its also used in the task_complete context.
