+++ default/tests/shared/database/task/test_task.py
@@ -3,7 +3,8 @@ from default.tests.test_utils import testing_setup
 from shared.tests.test_utils import common_actions, data_mocking
 from shared.database.task.task_user import TaskUser
 from unittest.mock import patch
-from methods.task.task.task_update import Task_Update
+from shared.utils.task.task_update_manager import Task_Update
+
 import analytics
 import requests
